requirements.txt
alabaster==0.7.12
asgiref==3.4.1
Babel==2.9.1
biopython==1.76
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
Django<4.0
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
docutils==0.16
gunicorn==20.1.0
idna==2.10
imagesize==1.2.0
Jinja2==2.11.3
loguru
MarkupSafe==1.1.1
numpy==1.20.1
packaging==20.9
pandas==1.2.3
psycopg2==2.8.6
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
semver==2.13.0
six==1.15.0
snowballstemmer==2.1.0
sqlparse==0.4.2
tabulate==0.8.9
urllib3==1.26.6
whitenoise==5.2.0
